+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Fixed
+- Detect fragment defaults when column operations have additional options or share an alter block
+
 ## [0.1.10] - 2026-03-07
 ### Fixed
 - Check constraint with `validate: false` is now treated as safe